move common.glue_tasks.CopyAllFiles to feff.firetaks.glue_tasks.CopyFeffOutputs
